requirements.txt
Django
requests
python-telegram-bot
openai